Added some initial setup
